move BIDSLayout.get() logic to query.py module and hook it into Folder class
